Move delete button to footer
